+++ src/gradient/Gradient.ts
