+++ src/shapes/Object/Object.ts
@@ -35,7 +35,7 @@ import {
   fabricObjectDefaultValues,
   stateProperties,
 } from './defaultValues';
-import type { Gradient } from '../../gradient/gradient.class';
+import type { Gradient } from '../../gradient/Gradient';
 import type { Pattern } from '../../pattern.class';
 import type { Canvas } from '../../canvas/canvas_events';
 import { removeTransformMatrixForSvgParsing } from '../../util/transform_matrix_removal';
