+++ src/typedefs.ts
@@ -1,6 +1,6 @@
 // https://www.typescriptlang.org/docs/handbook/utility-types.html
 import { BaseFabricObject } from './EventTypeDefs';
-import type { Gradient } from './gradient/gradient.class';
+import type { Gradient } from './gradient/Gradient';
 import type { Pattern } from './pattern.class';
 import type { Point } from './point.class';
 import type { FabricObject } from './shapes/Object/FabricObject';
